Move database and similar packages to components
This change moves and rearranges some our packages so they are easy to reference. The new `components` folder is for shared components that are used in all of our code. Previously these components were part of UCP which doesn't make sense, they are used by all of our control-plane services.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>