Remove ETCd database and secret store
These are unused now that we're using the simpler in-memory components for testing. Removing these simplifies the code and reduces our binary size by removing a chunky dependency.

For reviewers, the ETCd support being removed is NOT what users use today. That's the 'apiserver' database and the 'kubernetes' secret store. There is no change in behavior in this PR, we're removing something unused.

We're planning to use PostgreSQL for production-quality support in Radius, so we have no plans to use ETCd going forward.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>