Remove ETCd database and secret store (#8158)
# Description

These are unused now that we're using the simpler in-memory components
for testing. Removing these simplifies the code and reduces our binary
size by removing a chunky dependency.

For reviewers, the ETCd support being removed is NOT what users use
today. That's the 'apiserver' database and the 'kubernetes' secret
store. There is no change in behavior in this PR, we're removing
something unused.

We're planning to use PostgreSQL for production-quality support in
Radius, so we have no plans to use ETCd going forward.


## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).


## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>