Move shared services and host to components
This change moves the shared host package and diagnostics services to the new `pkg/components` folder. Also simplifies and improves consistency.

As part of this work, I made changes to the configuration file for consistency. Each of these diagnostics features now has a dedicated on/off switch. Where a service name is required, it is now specific in configuration rather than hardcoded in the code.

Also made the names of configuration files consistent and reflective of their scenario. All of the configuration files used for development reflect the 'dev' terminology. All of the configuration files used for self-hosted Radius (in Kubernetes) reflect the 'self-hosted' terminology.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>