Move shared services and host to components (#8160)
# Description

This change moves the shared host package and diagnostics services to
the new `pkg/components` folder. Also simplifies and improves
consistency.

As part of this work, I made changes to the configuration file for
consistency. Each of these diagnostics features now has a dedicated
on/off switch. Where a service name is required, it is now specific in
configuration rather than hardcoded in the code.

Also made the names of configuration files consistent and reflective of
their scenario. All of the configuration files used for development
reflect the 'dev' terminology. All of the configuration files used for
self-hosted Radius (in Kubernetes) reflect the 'self-hosted'
terminology.

## Type of change


- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).



## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>