Implement resource provider API for dynamic rp
This change implements the standard resource provider API for dynamic resources. This enables the dynamic-rp to serve the standard CRUDL API functionality for user-defined types.

This change does NOT add recipe support, that will come in the next change. The pattern that's supported by this change is an "inert" resource which has no backend functionality. Recipe support will be a superset of this functionality.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>