Implement resource provider API for dynamic rp (#8177)
# Description

This change implements the standard resource provider API for dynamic
resources. This enables the dynamic-rp to serve the standard CRUDL API
functionality for user-defined types.

This change does NOT add recipe support, that will come in the next
change. The pattern that's supported by this change is an "inert"
resource which has no backend functionality. Recipe support will be a
superset of this functionality.


## Type of change


Part of: #6688 

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>