Add capabilities to resource type API (#8182)
# Description

This change adds the 'capabilities' concept to the resource type API.

- Capabilities enable resource types to indicate the schema and
behaviors they support.
- Capabilities enable clients like the `rad` CLI to understand the
behaviors of resource types dynamically.

For example, we're adding the `SupportsRecipes` capability.

- All resource types that support recipes should declare this
capability. This is how a UDT will opt-in to recipe functionality during
provisioning.
- The `rad` CLI functionality for `rad recipe register` can introspect
the resource type to validate recipe support, rather than hardcoding
which types have the support and which don't.

----

Description of the changes:

- The manifests previously supported capabilities as part of the API
version, we're moving this to the resource type for a simplification.
- The manifest entry for capabilities wasn't sent to the server. Now it
is.
- Updated API, converters, and UCP functionality.


## Type of change

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).


Part of: #6688

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>