Fix for picking the right helm chart version during release cut (#8196)
# Description

To find the correct helm chart version in our code, we search for the
release major value but wildcard the patch value. We need to update the
versioning logic during the Helm install to handle a few different
cases:
1. If there is a pre-release version (i.e. -rc1, -rc2, etc), we want to
find a helm chart version that is an exact match
2. If there is a full release version, we want to find a helm chart
version that matches the release value but wildcards the patch value (in
case we have a patch release)
3. For dev builds, the release version will always be `0.42.42-dev` so
this value should be hardcoded as the version.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request fixes a bug in Radius and has an approved issue
(issue link required).
- This pull request adds or changes features of Radius and has an
approved issue (issue link required).
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: https://github.com/radius-project/radius/issues/8195

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

---------

Signed-off-by: Vishwanath Hiremath <vhiremath@microsoft.com>
(cherry picked from commit 5832dea7c1530cc88dd7d0730f3012c677393d85)