Implement fix for failures caused in test logging (#8190)
# Description

Logging in tests has an intentional limitation. Any logging performed
after the end of a test will cause the test to fail.

Unfortunately for us, this is a really hard limitation. We do a lot of
asynchronous processing, and we can't always guarantee that processing
is complete when a test completes. See
https://github.com/radius-project/radius/actions/runs/12591118184/job/35093755888
for an example failure. This is especially bad because it's never a
consistent repro.

This approach is based on the solution that Kubernetes uses:
https://github.com/kubernetes/klog/pull/337/files#diff-9c4e6a71ada6fa5d0ee80e05296e5b418351025a21b2f9dce9595da245272313
. They wrote their own logging package, so they can make the fix inside
of it. We're using logr (which was created by the same author), so we
have to make the fix outside of the the logging package, but it's the
same fundamental idea.



## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).


## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Ryan Nowak <nowakra@gmail.com>