Update instructions and config for required local dotnet version (#8213)
# Description

DE was updated to .net 8 in
https://github.com/radius-project/deployment-engine/pull/394. Similar
version is needed to run control plane locally, updating instructions
and `launch.json` to reflect that as the setup is broken with current
instructions due to version mismatch.

Validated updated config locally.

## Type of change

<!--

Please select **one** of the following options that describes your
change and delete the others. Clearly identifying the type of change you
are making will help us review your PR faster, and is used in authoring
release notes.

If you are making a bug fix or functionality change to Radius and do not
have an associated issue link please create one now.

-->

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

<!--

Please update the following to link the associated issue. This is
required for some kinds of changes (see above).

-->

Fixes: #issue_number

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [N/A] An overview of proposed schema changes is included in a linked
GitHub issue.
- [N/A] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [N/A] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [N/A] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [N/A] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [N/A] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Karishma Chawla <kachawla@microsoft.com>