Update Terraform Azure Test Resource Type (#8254)
# Description

We have been seeing flakyness in terraform recipe azure storage test due
to fluctuating create latencies of azure storage account provisioning
causing subsequent requests on the resource to fail due to 409. Since
the main goal of this test is to validate that Azure provider/auth is
correctly configured for Terraform recipes, we don't specifically need
to provision Azure storage. So switching the test to create a resource
group, which will provision faster for reliability in testing.

More details and investigation on this are logged in the comments here:
https://github.com/radius-project/radius/issues/8237.

## Type of change

Fixes: #8172

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [NA] An overview of proposed schema changes is included in a linked
GitHub issue.
- [NA] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [NA] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [NA] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [NA] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [NA] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

Signed-off-by: Karishma Chawla <kachawla@microsoft.com>