Set cache option on setup-go action in workflows (#8280)
# Description

This PR sets the `cache: true` option on the `setup-go` actions in the
workflows. Previously we used a common caching technique to cache the go
install and modules folders, but this feature is now built into the
[`setup-go`
action](https://github.com/actions/setup-go/blob/main/docs/adrs/0000-caching-dependencies.md).
This change was tested with PR #8276. Following that, this PR updates
the remaining workflows to use the new caching feature.

The default value for the `cache` option is `true` on the `setup-go@v5`
action. This PR explicitly sets the value to `true` on the action where
it is not specified so that the option will be obvious, even though it
will not change the behavior of the action for cases when the option was
not set.

- `build.yaml`: Explicitly set cache option to true, which is the same
as the default value.
- `functional-test-cloud.yaml`: Changed the cache option from false to
true and removed the original caching actions.
- `lint.yaml`: Explicitly set cache option to true, which is the same as
the default value.
- `long-running-azure.yaml`: Explicitly set cache option to true, which
is the same as the default value.
- `publish-docs.yaml`: Explicitly set cache option to true, which is the
same as the default value.

> NOTE: This PR is created from a branch in this repo instead of a fork
from an external repo to allow us to run the cloud test workflows.

## Type of change

- This pull request adds or changes features of Radius and has an
approved issue (issue link required).

Related to issue: #8274
Related to PR #8276

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- [ ] An overview of proposed schema changes is included in a linked
GitHub issue.
- [ ] A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
- [ ] If applicable, design document has been reviewed and approved by
Radius maintainers/approvers.
- [ ] A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
- [ ] A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
- [ ] A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.

---------

Signed-off-by: Brooke Hamilton <45323234+brooke-hamilton@users.noreply.github.com>
Signed-off-by: Nick Beenham <Nicholas_Beenham@cable.comcast.com>