Add maintainer/approver groups as dependency file owners (#8933)
# Description
Two updates

* Add maintainer/approver groups as dependency file owners: The most
specific rule takes precedence over more general ones. This means that
if a file or directory matches multiple rules, the one with the narrower
scope (e.g., a specific file or subdirectory) will override more general
ones (e.g., the root * rule). So we need to add maintainer/approver
teams explicitly to the files that oncall team is added as owner for.

* Add oncall team as owner for dockerfile for snyk updates

---------

Signed-off-by: Karishma Chawla <kachawla@microsoft.com>