Updating launch.json and running-controlplane-locally doc for updated DE (#8932)
# Description

### **Problem**

The **_bicep-extensibility_** library was updated to require an encoded
kubeConfig string in the request payload:

[ExtensibilityOperationRequestExtensions.cs#L66](https://github.com/Azure/bicep-extensibility/blob/main/src/Azure.Deployments.Extensibility.Providers.Kubernetes/Extensions/ExtensibilityOperationRequestExtensions.cs#L66)

However, in our current recipes, we set **kubeConfig** to an empty
string:
[Search
reference](https://github.com/search?q=repo%3Aradius-project%2Fradius%20%22kubeConfig%3A%22&type=code)

This results in the extensibility provider not having the required
Kubernetes config context to execute operations. We need to ensure that
kubeConfig is properly set in the request after the Deployment Engine
(DE) receives it from UCP.

### **Solution**

We handle this differently depending on the runtime environment:

**Production scenarios (DE running in a cluster):**

Use Kubernetes in-cluster configuration via
KubernetesClientConfiguration.InClusterConfig:

https://github.com/kubernetes-client/csharp/blob/master/src/KubernetesClient/KubernetesClientConfiguration.InCluster.cs#L41

**Local development scenarios (DE running outside a cluster):**

* Load **kubeConfig** from the user’s local kubeconfig file.
* If the active Radius workspace does **not** match the current
Kubernetes context in the kubeconfig, we throw an error to avoid
accidental mis-targeting.
* This is to ensure that the local development environment is configured
intentionally and predictably.

### **Implementation Notes**

We introduce the following environment variables to support this logic:

* **RAD_LOCAL_DEVELOPMENT** (required):

Must be set to true when running the control plane locally. This flag
determines whether to use local vs. in-cluster kubeconfig resolution.

* **KUBECONFIG** (optional):

Can be set to override the default kubeconfig file location
(~/.kube/config).

## Type of change
- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: #issue_number

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [ ] Yes
    - [x] Not applicable
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [ ] Yes
    - [x] Not applicable
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [ ] Yes
    - [x] Not applicable

Signed-off-by: ytimocin <ytimocin@microsoft.com>